requirement.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
exceptiongroup==1.2.2
FLAML==2.1.2
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
kiwisolver==1.4.5
matplotlib==3.8.4
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.20.0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pyautogen==0.2.26
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1